+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Restore previous windows dimensions with next start of ASGDX [#1879](https://github.com/thorchain/asgardex-electron/issues/1879)
 - Show asset icon in TxDetail (wallet history + pool details)[#1955](https://github.com/thorchain/asgardex-electron/pull/1955)
 - [ADD] Check asset mismatch for Ledger + keystore [#1938](https://github.com/thorchain/asgardex-electron/issues/1938)
+- Show ledger | keystore addresses in tooltips [#1959](https://github.com/thorchain/asgardex-electron/pull/1959)
 
 ## Update